+++ src/it/java/org/greencodeinitiative/creedengo/python/integration/tests/GCIRulesIT.java
@@ -273,4 +273,48 @@ class GCIRulesIT extends GCIRulesBase {
 
     }
 
+    @Test
+    void testGCI96() {
+        String filePath = "src/pandasRequireUsecols.py";
+        String ruleId = "creedengo-python:GCI96";
+        String ruleMsg = "Specify 'usecols' or 'columns' when reading a DataFrame using Pandas to load only necessary columns";
+        int[] startLines = new int[]{
+            3, 4, 5, 6, 7, 16, 19
+        };
+        int[] endLines = new int[]{
+            3, 4, 5, 6, 7, 16, 19
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_10MIN);
+    }
+    void testGCI97(){
+        String filePath = "src/optimizeSquareComputation.py";
+        String ruleId = "creedengo-python:GCI97";
+        String ruleMsg = "Use x*x instead of x**2 or math.pow(x,2) to calculate the square of a value";
+        int[] startLines = new int[]{
+            4, 7, 19, 20, 25, 26, 31, 38
+        };
+        int[] endLines = new int[]{
+            4, 7, 19, 20, 25, 26, 31, 38
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1MIN);
+    }
+    
+    void testGCI103(){
+
+        String filePath = "src/dictionaryItemsUnused.py";
+        String ruleId = "creedengo-python:GCI103";
+        String ruleMsg = "Use dict.keys() or dict.values() instead of dict.items() when only one part of the key-value pair is used";
+        int[] startLines = new int[]{
+            5, 8, 12, 32, 35, 44
+        };
+        int[] endLines = new int[]{
+            5, 8, 12, 32, 35, 44
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1MIN);  
+
+    }
+
 }